# spec/swagger.yaml
definitions:
...
  handlers.DailyTokenUsage:
    properties:
      cache_read_tokens:
        type: integer
      day:
        type: string
      input_tokens:
        type: integer
      output_tokens:
        type: integer
      reasoning_tokens:
        type: integer
    type: object
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.ModelTokenUsage:
    properties:
      input_tokens:
        type: integer
      model_id:
        type: string
      model_name:
        type: string
      model_slug:
        type: string
      output_tokens:
        type: integer
      provider_name:
        type: string
    type: object
...
  handlers.TokenUsageRecord:
    properties:
      cache_read_tokens:
        type: integer
      created_at:
        type: string
      id:
        type: string
      input_tokens:
        type: integer
      model_id:
        type: string
      model_name:
        type: string
      model_slug:
        type: string
      output_tokens:
        type: integer
      provider_name:
        type: string
      reasoning_tokens:
        type: integer
      session_id:
        type: string
      session_type:
        type: string
    type: object
  handlers.TokenUsageRecordsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.TokenUsageRecord'
        type: array
      total:
        type: integer
    type: object
  handlers.TokenUsageResponse:
    properties:
      by_model:
        items:
          $ref: '#/definitions/handlers.ModelTokenUsage'
        type: array
      chat:
        items:
          $ref: '#/definitions/handlers.DailyTokenUsage'
        type: array
      heartbeat:
        items:
          $ref: '#/definitions/handlers.DailyTokenUsage'
        type: array
      schedule:
        items:
          $ref: '#/definitions/handlers.DailyTokenUsage'
        type: array
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /bots/{bot_id}/token-usage:
    get:
      description: Get daily aggregated token usage for a bot, split by chat, heartbeat,
        and schedule session types, with optional model filter and per-model breakdown
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        required: true
        type: string
      - description: End date exclusive (YYYY-MM-DD)
        in: query
        name: to
        required: true
        type: string
      - description: Optional model UUID to filter by
        in: query
        name: model_id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TokenUsageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get token usage statistics
      tags:
      - token-usage
  /bots/{bot_id}/token-usage/records:
    get:
      description: Paginated list of individual LLM call records (assistant messages
        with usage) for a bot, with optional model and session type filters
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        required: true
        type: string
      - description: End date exclusive (YYYY-MM-DD)
        in: query
        name: to
        required: true
        type: string
      - description: Optional model UUID to filter by
        in: query
        name: model_id
        type: string
      - description: 'Optional session type: chat, heartbeat, schedule, or acp_agent'
        in: query
        name: session_type
        type: string
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TokenUsageRecordsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List per-call token usage records
      tags:
      - token-usage
...
swagger: "2.0"